+++ docs/Changelog.md
@@ -4,6 +4,7 @@
   - fix crashing on second start jmeter under GUI mode
   - iterate requests forever if no limits were specified
   - fix test duration logic to respect iterations set
+  - distributed tests are now supported with JMeter UI (gui: true option)
 
 # 0.2.23
   - rename `bulk-size` option into `send-interval` for BlazeMeter reporter